Linux fix. column doesnt support colorized input
Mac OS X's column and Linux's column are different. Mac OS X column (or
FreeBSD column I think) ignore escape codes, Linux doesn't so the output
is wront when the input is colorized. The script detects if Linux or Mac
OS X and reverts to no color if in Linux